Started resizing image to screen size. Not finished, check trello for inconsistent exception description
